revision 13128 - dashboard voxel human ik rig
i wanted to make the voxel human ik rig walk on the voxel virtual desktop surface but that is more complicated to code, and it had been a while that i hadn't touched inverse kinematics so i decided to practice by re-incorporating my ik development and tested for camera corners "dashboard figurines" and after a couple days of work i have a result.

But my development for trying to fix the virtual desktop screencapture image when using an NVIDIA video card didn't work. I tried playing around with the samplerstate from within the shader and outside of the shader, i tried padding completely my vertex binding in divisions of 4 by adding two more vertex bindings for an int of 4 bytes each and also did some other tests. In the end, it seems to be maybe a projection matrix/viewmatrix issues, maybe it's a translation of the texture that i am doing on the pixel shader that is accepted with AMD video cards but isn't accepted on NVIDIA video cards. It's maybe one of those things.

steve chassé