rev13131 - added voxel ik head. bigger dashboard figurines.
rev13131 - i added back the voxel ik head on the voxel ik human rigs. I made the dashboard figurines bigger and now the hands at least are aligned with the lower arm but the hand is not rotating yet towards the cursor location but the hand is clenching and the index is pointing the finger when left mouse clicking.

steve chassé 9