revision 13135 - calibration update - halfway there
hey everyone, i was able to calibrate approx 75 resolutions on 150. It's all of the low voxel definition resolutions that i worked on but out the 75 resolutions i calibrated, i will review 13 of those as i am unsure, and wrote next to the resolutions "to review" in the script scgraphicssec.cs.

steve chassé aka 9