revision 13136 - i removed the oculus rift library
i had completely forgotten to remove the oculus rift library and the code associated with it. But in a rush tonight also for working on providing a document containing 30 pages of basic information on my program, i corrected a few things like a menu that wasn't hidding and added extra words next to the options #1 to #5. I moved the menu notes at the top left since it was hiding a corner often used for fps display.