update - revision 1.31371 - fix on vvd idle state heightmaps not resetting.
the heightmap values weren't resetting when closing a captured program. Now the heightmaps values are correctly resetting. My tests yesterday night were showing that replacing the captured image by a null image, was actually resetting the heightmaps values as black color values were sent to the shader. But this morning that didn't work so i manually setup some code in order to directly reset the heightmaps values if a captured program is closed.

steve chassé